+++ CHANGELOG.md
@@ -5,6 +5,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](https://semver.org/).
 
 ## [Unreleased]
+- Add a feature to use a template file for `add` command. (#87)
 - Add new keywords for `list` command. (#90)
   - `this_month` and `last_month`
 - Add a new option, `verbose` for `list` command. (#76)